Fix relative path
