+++ RuntimeInternals/AssemblyInfo.cs
@@ -9,4 +9,5 @@ using System.Runtime.CompilerServices;
     @"This assembly can be used from the runtime code because it does not depend on test-framework.
 This assembly is named ""Internal"", however, the included classes are public.")]
 
+[assembly: InternalsVisibleTo("TestHelper.Editor")]
 [assembly: InternalsVisibleTo("TestHelper.RuntimeInternals.Tests")]
